+++ safesafe/Services/Free test/Model/JS/FreeTestPinCodeResponse.swift
@@ -1,0 +1,12 @@
+//
+//  FreeTestPinCodeResponse.swift
+//  safesafe
+//
+//  Created by Lukasz szyszkowski on 23/10/2020.
+//
+
+import Foundation
+
+struct FreeTestPinCodeResponse: Codable {
+    let code: String?
+}