+++ safesafe/Services/Free test/Model/Networking/FreeTestCreateSubscriptionRequestModel.swift
@@ -1,0 +1,13 @@
+//
+//  FreeTestCreateSubscriptionRequestModel.swift
+//  safesafe
+//
+//  Created by Lukasz szyszkowski on 22/10/2020.
+//
+
+import Foundation
+
+struct FreeTestCreateSubscriptionRequestModel: Codable {
+    let code: String
+    let guid: String
+}